image_compress/requirements.txt
pillow
flask